refactor(breakout): decompose god-object command handler and standardize subcommand options
- Extract subcommand handlers into modular files under src/modules/breakout/handlers/
- Standardize option naming for 'end' subcommand option to 'mainroom'
- Simplify commands/main/breakout.ts from 600 lines to a 160-line router table
